requirements.txt
discord.py
dash
gunicorn